Tweaks so we still compile
